Adding AbstractLogger implementation, moving NullLogger to subclass
